+++ x/leverage/client/tests/tests.go
@@ -328,11 +328,14 @@ func (s *IntegrationTests) TestLeverageScenario() {
 				// app/test_helpers.go/IntegrationTestNetworkConfig
 				// times the amount of umee, and then times params
 				// (1001 / 1000000) * 34.21 = 0.03424421
-				SuppliedValue: sdk.MustNewDecFromStr("0.03424421"),
+				SuppliedValue:     sdk.MustNewDecFromStr("0.03424421"),
+				SpotSuppliedValue: sdk.MustNewDecFromStr("0.03424421"),
 				// (1001 / 1000000) * 34.21 = 0.03424421
-				CollateralValue: sdk.MustNewDecFromStr("0.03424421"),
+				CollateralValue:     sdk.MustNewDecFromStr("0.03424421"),
+				SpotCollateralValue: sdk.MustNewDecFromStr("0.03424421"),
 				// (251 / 1000000) * 34.21 = 0.00858671
-				BorrowedValue: sdk.MustNewDecFromStr("0.00858671"),
+				BorrowedValue:     sdk.MustNewDecFromStr("0.00858671"),
+				SpotBorrowedValue: sdk.MustNewDecFromStr("0.00858671"),
 				// (1001 / 1000000) * 34.21 * 0.25 = 0.0085610525
 				BorrowLimit: sdk.MustNewDecFromStr("0.0085610525"),
 				// (1001 / 1000000) * 0.26 * 34.21 = 0.008903494600000000