+++ CHANGELOG.md
@@ -44,12 +44,16 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
-## [Unreleased]
+## v6.1.0 - 2023-09-29
 
 ### Improvements
 
 - [2261](https://github.com/umee-network/umee/pull/2261) Use go 1.21
 - [2263](https://github.com/umee-network/umee/pull/2263) Add spot price fields to account summary, and ensure all other fields use leverage logic prices.
+
+### Features
+
+- [2269](https://github.com/umee-network/umee/pull/2269) Enable x/metoken module.
 
 ### Bug Fixes